refactor(flow): move helpers to reduce flow_status.py line count
Moved _parse_remote_issue_number() and _timeline_to_flow_events()
to flow_status_helpers.py to bring flow_status.py under 400-line
CI block threshold (now 395 lines).

CI was failing with:
- flow_status.py: 426 lines (exceeds 400-line limit)

After refactoring:
- flow_status.py: 395 lines ✓
- flow_status_helpers.py: 202 lines

All 6 remote timeline tests still pass.
